+++ DIOPI-ADAPTOR/codegen/op_template.py
@@ -14,9 +14,102 @@ class OpTemplate(object):
 #include <vector>
 #include <diopi/diopirt.h>
 #include <diopi/functions.h>
-#include <utils.hpp>
 
 namespace diopiadaptor{
+
+inline std::vector<int64_t> calcStrides(int ndims, diopiSize_t size, diopiMemoryFormat_t format = diopiMemoryFormat_t::Contiguous) {
+    std::vector<int64_t> strides;
+    strides.resize(ndims);
+    int64_t st = 1;
+    if (format == diopiMemoryFormat_t::Contiguous) {
+        for (int64_t i = ndims; i > 0; --i) {
+            strides[i - 1] = st;
+            if (size.data[i - 1] == 0) continue;
+            if (size.data[i - 1] == -1) st = -1;
+            if (st != -1) st *= size.data[i - 1];
+        }
+    } else if (format == diopiMemoryFormat_t::ChannelsLast) {
+        for (auto k : {1, 3, 2, 0}) {
+            strides[k] = st;
+            if (size.data[k] == 0) {
+                continue;
+            }
+            if (size.data[k] == -1) st = -1;
+            if (st != -1) st *= size.data[k];
+        }
+    } else if (format == diopiMemoryFormat_t::ChannelsLast3d) {
+        for (auto k : {1, 4, 3, 2, 0}) {
+            strides[k] = st;
+            if (size.data[k] == 0) {
+                continue;
+            }
+            if (size.data[k] == -1) st = -1;
+            if (st != -1) {
+                st *= size.data[k];
+            }
+        }
+
+    } 
+    else {
+        // PARROTS_THROW(InvalidArgs) <<
+        //         "Invalid MemoryFormat " << memoryFormatName(format);
+    }
+    return strides;
+}
+
+inline bool isLikeChannelsLast(diopiConstTensorHandle_t tensor, bool checkContiguous, diopiMemoryFormat_t format = diopiMemoryFormat_t::ChannelsLast) {
+    diopiSize_t shape, stride;
+    diopiGetTensorShape(tensor, &shape);
+    diopiGetTensorStride(tensor, &stride);
+    if (shape.len != 4) return false;
+    int64_t totalSize = 1;
+    for (int64_t i = 0; i < shape.len; ++i) {
+        totalSize *= shape.data[i];
+    }
+    if (totalSize == 0) return false;
+    if (stride.data[0] == stride.data[1]) return false;
+    if (checkContiguous) {
+        auto realStride = calcStrides(shape.len, shape, format);
+        for (int i = 0; i < stride.len; ++i) {
+            if (i >= realStride.size() || realStride[i] != stride.data[i]) {
+                return false;
+            }
+        }
+        return true;
+    } else {
+        int64_t st = 1;
+        std::vector<int> orders;
+        if (format == diopiMemoryFormat_t::ChannelsLast)
+            orders = {1, 3, 2, 0};
+        else if (format == diopiMemoryFormat_t::ChannelsLast3d)
+            orders = {1, 4, 3, 2, 0};
+        for (auto k : orders) {
+            if (stride.data[k] < st) return false;
+            st = stride.data[k] * shape.data[k];
+        }
+        return true;
+    }
+}
+
+inline diopiMemoryFormat_t probableMemoryFormat(diopiConstTensorHandle_t tensor, bool exactMatch = false) {
+    return isLikeChannelsLast(tensor, exactMatch) ? diopiMemoryFormat_t::ChannelsLast
+        : (isLikeChannelsLast(tensor, exactMatch, diopiMemoryFormat_t::ChannelsLast3d) ? diopiMemoryFormat_t::ChannelsLast3d
+        : diopiMemoryFormat_t::Contiguous);
+}
+
+inline bool isContiguous(diopiSize_t size, diopiSize_t stride, diopiMemoryFormat_t format = diopiMemoryFormat_t::Contiguous) {
+    int64_t totalSize = 1;
+    for (int64_t i = 0; i < size.len; ++i) {
+        totalSize *= size.data[i];
+    }
+    if (totalSize == 0) return false;
+    if (format == diopiMemoryFormat_t::ChannelsLast && size.len != 4) return false;
+    auto st = calcStrides(size.len, size, format);
+    for (int64_t i = 0; i < size.len; ++i) {
+        if (st[i] != stride.data[i] && size.data[i] > 1) return false;
+    }
+    return true;
+}
 
 static std::vector<diopiMemoryFormat_t> defaultFormats{diopiMemoryFormat_t::Contiguous, diopiMemoryFormat_t::ChannelsLast};
